examples/speech_commands_recognition_with_llm/server/requirements.txt
pyserial>=3.5
pydub>=0.25.1
wave